move naive_assert to own file
