+++ src/style-spec/expression/definitions/coalesce.ts
@@ -1,4 +1,4 @@
-import {naiveAssert} from '../../../util/test/util';
+import {naiveAssert} from '../../../util/test/naive_assert';
 
 import {checkSubtype, ValueType} from '../types';
 import ResolvedImage from '../types/resolved_image';
