+++ src/ui/camera.ts
@@ -5,7 +5,7 @@ import LngLat from '../geo/lng_lat';
 import LngLatBounds from '../geo/lng_lat_bounds';
 import Point from '@mapbox/point-geometry';
 import {Event, Evented} from '../util/evented';
-import {naiveAssert} from '../util/test/util';
+import {naiveAssert} from '../util/test/naive_assert';
 import {Debug} from '../util/debug';
 
 import type Transform from '../geo/transform';
