+++ src/util/dictionary_coder.ts
@@ -1,4 +1,4 @@
-import {naiveAssert} from '../util/test/util';
+import {naiveAssert} from '../util/test/naive_assert';
 
 class DictionaryCoder {
     _stringToNumber: {[_: string]: number};
