+++ src/util/dom.ts
@@ -1,6 +1,6 @@
 import Point from '@mapbox/point-geometry';
 
-import {naiveAssert} from '../util/test/util';
+import {naiveAssert} from '../util/test/naive_assert';
 
 export default class DOM {
     private static readonly docStyle = typeof window !== 'undefined' && window.document && window.document.documentElement.style;
