+++ src/util/test/naive_assert.ts
@@ -1,0 +1,3 @@
+export function naiveAssert(test: unknown, message?: string) {
+    if (!test) throw new Error(message.toString());
+}
