+++ test/bench/benchmarks/filter_evaluate.ts
@@ -4,7 +4,7 @@ import {VectorTile} from '@mapbox/vector-tile';
 import Pbf from 'pbf';
 import createFilter from '../../../src/style-spec/feature_filter';
 import filters from '../data/filters.json';
-import {naiveAssert} from '../../../src/util/test/util';
+import {naiveAssert} from '../../../src/util/test/naive_assert';
 
 export default class FilterEvaluate extends Benchmark {
     layers: any[];
